+++ src/main/java/com/ronlab/rga/party/Party.java
@@ -6,7 +6,7 @@ import java.util.*;
 
 public class Party {
 
-    public enum State { LOBBY, IN_GAME }
+    public enum State { LOBBY, QUEUED, IN_GAME }
 
     private final UUID id;
     private final String minigameId;